now can upload gameData (backend testing)

+++ FrontEnd/node/buffer.js
@@ -9,15 +9,18 @@ game.stdout.on('data',(data)=>{
     if(canOutput){
         console.log(`stdout:${data.toString()}`);
         fs.writeFile(`${__dirname}/exe/stdout.txt`,data.toString().replaceAll(/[\u0000-\u001F\u007F-\u009F]/g,''),()=>{});
+        fs.writeFile(`${__dirname}/exe/stdin.txt`,"",()=>{});
     }
     canOutput=false;
 })
 
 console.log(`game loaded!`);
+fs.writeFile(`${__dirname}/exe/stdin.txt`,"",()=>{});
 
 let interval=setInterval(()=>{
     fs.readFile(`${__dirname}/exe/stdin.txt`,'utf-8',(err,data)=>{
         if(err) return;
+        if(!data.toString().replaceAll(/[\u0000-\u001F\u007F-\u009F]/g,'').length) return;
         fs.unlink(`${__dirname}/exe/stdin.txt`,()=>{});
         canOutput=true;
         data.toString().replaceAll(/[\u0000-\u001F\u007F-\u009F]/g,'').split(' ').forEach((command)=>{
